fix: simplify root index
