+++ frontend/src/index.tsx
@@ -1,18 +1,12 @@
 import './init';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { ClassPrefix } from '@carbon/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { isAxiosError } from 'axios';
 
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { ThemePreference } from './utils/ThemePreference';
 import { env } from './env';
 import makeServer from './mock-server/server';
 import AuthProvider from './contexts/AuthProvider';
-import prefix from './styles/classPrefix';
 
 const appVersion: string = env.VITE_NRSPARWEBAPP_VERSION || 'dev';
 
@@ -24,33 +18,6 @@ if (isDevEnv) {
   makeServer('development');
 }
 
-const HTTP_STATUS_TO_NOT_RETRY = [400, 401, 403, 404];
-const MAX_RETRIES = 3;
-
-const queryClient = new QueryClient(
-  {
-    defaultOptions: {
-      queries: {
-        refetchOnMount: false,
-        refetchOnWindowFocus: false,
-        // Do not retry on errors defined above
-        retry: (failureCount, error) => {
-          if (failureCount > MAX_RETRIES) {
-            return false;
-          }
-          if (
-            isAxiosError(error)
-            && HTTP_STATUS_TO_NOT_RETRY.includes(error.response?.status ?? 0)
-          ) {
-            return false;
-          }
-          return true;
-        }
-      }
-    }
-  }
-);
-
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
@@ -58,16 +25,7 @@ const root = ReactDOM.createRoot(
 root.render(
   <AuthProvider>
     <React.StrictMode>
-      <ClassPrefix prefix={prefix}>
-        <ThemePreference>
-          <QueryClientProvider client={queryClient}>
-            <App />
-            {
-              isDevEnv && <ReactQueryDevtools initialIsOpen={false} />
-            }
-          </QueryClientProvider>
-        </ThemePreference>
-      </ClassPrefix>
+      <App />
     </React.StrictMode>
   </AuthProvider>
 );
